==> Codes/02_test_cospectral_raw_sim100.Rmd <==
---
title: "wavelet angle test"
author: "Moeka"
date: "2025-05-13"
output: github_document
---
wavelet analysis: chrome-extension:http://www.hs-stat.com/projects/WaveletComp/WaveletComp_guided_tour.pdf

- Normalization
All time series should be normalized to have zero mean and unit variance prior to wavelet analysis.  

- Gaps
Missing values in all the variables should be padded with zeros.


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(WaveletComp)
library(lubridate)
library(dplyr)
library(ggplot2)
library(ggpubr)
library(tidyr)
library(broom)
library(forestmangr)
library(gridExtra)
library(purrr)
```


```{r read df}
setwd("G:/My Drive/Research/Projects/CRK_auto/Data/GPP_SR_combined")

GPP_SR_MJ22 <- read.csv("GPP_SR_MJ22.csv") %>% mutate(date = as.POSIXct(Datetime_hr_CST, tz = "Etc/GMT+6"))
GPP_SR_MA23 <- read.csv("GPP_SR_MA23.csv") %>% mutate(date = as.POSIXct(Datetime_hr_CST, tz = "Etc/GMT+6"))
GPP_SR_Mar24 <- read.csv("GPP_SR_Mar24.csv") %>% mutate(date = as.POSIXct(Datetime_hr_CST, tz = "Etc/GMT+6"))
GPP_SR_Apr24 <- read.csv("GPP_SR_Apr24.csv") %>% mutate(date = as.POSIXct(Datetime_hr_CST, tz = "Etc/GMT+6"))
GPP_SR_Sep24 <- read.csv("GPP_SR_Sep24.csv") %>% mutate(date = as.POSIXct(Datetime_hr_CST, tz = "Etc/GMT+6"))
GPP_SR_Oct24 <- read.csv("GPP_SR_Oct24.csv") %>% mutate(date = as.POSIXct(Datetime_hr_CST, tz = "Etc/GMT+6"))
```

```{r Prep before the wavelet analysis}
# normalize with 0 mean and unit variance - use scale()

# for cross wavelet analysis with Morlet function (continuous) 
scale_pad <- function(df, var1, var2){
  df %>%
    select(date, var1, var2) %>%
    mutate(NA_flag = ifelse(!is.na(df[[var1]]) & !is.na(df[[var2]]), "0", "1")) %>%
    mutate(across(where(is.numeric), scale)) %>%
    mutate(across(where(is.numeric), ~replace(., is.na(.), 0)))
}

# function to check the amount (%) of the gaps
pMiss <- function(x){
  sum(is.na(x)) / length(x) * 100
  }


# Detect gap lengths in the time series
detect_gaps <- function(data, col) {
  
  mt <- is.na(data[, col])
  ind <- 1 # index for marking the gaps
  mk <- vector()
  
  for (i in 1:length(mt)) {
    if (mt[i] == FALSE) {
      mk[i] <- 0 # non-gaps are marked as 0
    } else {
      if (mt[i] == TRUE) {
        mk[i] <- ind # gaps are marked as the value of ind
        if (i != length(mt)){ # to prevent the error when loop reach the end
          if (mt[i + 1] == FALSE) {
            ind <- ind + 1 # when reached the end of a gap, change add 1 to ind
          }
        }
      }
    }
  }
  
  stat <- table(mk)[-1] # number of data points in each gap
  # print using "cat" for break into lines
  cat(paste0("","\n",
             "##### Summary #####","\n",
             "","\n",
             "Total gaps:       ",max(mk),"\n",
             "< 6 hrs:          ",sum(stat < 6),"\n",
             ">= 6 hrs < 1 day: ",sum(stat >= 6 & stat < 24),"\n",
             ">= 1 & < 2 days:  ",sum(stat >= 24 & stat < 24*2),"\n",
             ">= 2 days:        ",sum(stat >= 24*2)
  ))
}


# function to write a list of dfs to csvs
save_to_csv <- function(list_of_dfs, dir, fname){
  for (i in seq_along(list_of_dfs)) {
  write.csv(list_of_dfs[[i]], paste0(dir, fname, names(list_of_dfs)[i], ".csv"), row.names = FALSE)
  }
}

```

```{r cross wavelet - continuous}
coh <- function(df, var1, var2, upper){
  df1 <- scale_pad(df, var1, var2)
  analyze.coherency(df1, my.pair = c(var1, var2),
                          loess.span = 0,     ## no detrending required
                          dt = 1/24,          ## defines the time unit --- there are 24 slots in a day! Time unit = 1 day
                          dj = 1/100,         ## resolution along period axis
                          lowerPeriod = 1/4,  ## 6 hour
                          upperPeriod = upper,
                          make.pval = T,      ## check for significance
                          n.sim = 100)         ## 10 white-noise simulations used for assessing significance 
}

# function to generate a cross-wavelet power spectrum
wc_img <- 
  function(df, title){
  
  dev.new(width = 6, height = 5, noRStudioGD = TRUE)
  
  wc.image(
    df,
    legend.params = 
      list(
        lab = "cross-wavelet power levels", # axis label
        lab.line = 1.5, # distance between bar label and axis label
        mar = 5
        ), 
    lvl = 0.3,
    timelab = "",
    show.date = T, 
    date.format = "%Y-%m-%d %H:%M:%S",
    periodlab = "period (days)",
    main = title
    )

}

# function to analyze and plot them all not recommended for a longer period  of
# term

# does not work well when assign an object name..
# want to make a timestamp x axis - too crowded w/ hourly timestamps
wc_summary <- function(df, resp_var){
  
  wc_results <- list()
  
  # wc result
  wc_results[["GPP"]] <- coh(df, resp_var, "GPP", 64)
  wc_results[["PAR"]] <- coh(df, resp_var, "PPFD_IN_1_1_1", 64)
  wc_results[["Ts"]] <- coh(df, resp_var,"TS_1_1_1", 64)
  wc_results[["VWC"]] <- coh(df, resp_var, "SWC_1_1_1", 64)

  return(wc_results)
  }

# this will generate all 4 plots one by one
wc_img_gen <- function(df, resp_var){
 # par(mar=c(2,2,1.5,0.3)+0.1)
  par(plt = c(0.1, 0.95, 0.1, 0.95))
  wc_img(df[["GPP"]], paste0(resp_var, " vs GPP"))
  wc_img(df[["PAR"]], paste0(resp_var, " vs PAR"))
  wc_img(df[["Ts"]], paste0(resp_var, " vs Ts"))
  wc_img(df[["VWC"]], paste0(resp_var, " vs VWC"))

}

# this will generate all 4 plots one by one
wc_avg_gen <- function(df, resp_var){
  
  wc.avg(df[["GPP"]], main = paste0(resp_var, " vs "), siglvl = 0.05, sigcol = "black", 
         sigpch = 20, periodlab = "", show.legend = F, averagelab = "", maximum.level = 25)
  
  wc.avg(df[["PAR"]], main = paste0(resp_var, " vs PAR"), siglvl = 0.05, sigcol = "black", 
         sigpch = 20, periodlab = "", show.legend = F, averagelab = "", maximum.level = 25)
  
  wc.avg(df[["Ts"]], main = paste0(resp_var, " vs Ts"), siglvl = 0.05, sigcol = "black", 
         sigpch = 20, periodlab = "", show.legend = F, averagelab = "", maximum.level = 25)
  
  wc.avg(df[["SWC"]], main = paste0(resp_var, " vs SWC"), siglvl = 0.05, sigcol = "black", 
         sigpch = 20, periodlab = "", show.legend = F, averagelab = "", maximum.level = 25)
  
}

# function to generate result df
coh_result <- function(pad_df, coh_df){
  
  # extracting the p-value at Period 1
  result <- 
    cbind(pad_df, data.frame(pval_period1 = coh_df$Power.xy.pval[201,])) 
  
  # Average angles between Period 0.5 - 1.5
  # ASSUMING THE Y_AXIS IS LOG-BASED
  result$Angle_raw <- colMeans(data.frame(coh_df$Angle[101:211,])) 
  result$Power_raw <- colMeans(data.frame(coh_df$Power.xy[101:211,])) 
  
  # Result df with NA_flag, and running averages
  # heatmap arrows are generated with p<0.05.
  # I think p = pval.xy
  result <- 
    result %>%
    mutate(Angle = ifelse(NA_flag == "0" & pval_period1 < 0.1, Angle_raw, NA),
           Angle_1d = zoo::rollmean(Angle, k = 24, fill = NA),      # 1 day moving avg - 24
           Angle_7d = zoo::rollmean(Angle, k = 24*7, fill = NA),    # 7 days moving avg - 24*7
           Angle_15d = zoo::rollmean(Angle, k = 24*15, fill = NA),  # 15 days moving avg - 24*15
           Power = ifelse(NA_flag == "0" & pval_period1 < 0.1, Power_raw, NA),
           Power_1d = zoo::rollmean(Power, k = 24, fill = NA),      # 1 day moving avg - 24
           Power_7d = zoo::rollmean(Power, k = 24*7, fill = NA),    # 7 days moving avg - 24*7
           Power_15d = zoo::rollmean(Power, k = 24*15, fill = NA)  # 15 days moving avg - 24*15
            )
    return(result)
}


# FOR VIZ
wc_lag_img <- function(result_df, title){
  
  result_df$date <- as_datetime(result_df$date)
  
  # 1: the midtime to have a time (00:00)
  # 2: if / included use mdy_hm, not use ymd_hms
  # result_df$date <- ifelse(grepl("/", result_df$date), 
  #                          as_datetime(ifelse(grepl(":",date), as.POSIXct(date, format = '%m/%d/%Y %H:%M'),
  #                                          as.POSIXct(date, format = '%m/%d/%Y'))),
  #                          as_datetime(date) 
  #                          )
  
  pivot_longer(result_df, 
               cols = c(Angle, Angle_1d, Angle_7d, Angle_15d),
               names_to = "Variable", 
               values_to = "Value"
               ) %>%
    ggplot() +
    geom_point(aes(date, 24 * Value / (2 * pi), col = Variable)) +
    scale_color_manual(breaks = c("Angle", "Angle_1d", "Angle_7d", "Angle_15d"),
                       values = c("#A0B1BA", "#045275", "#089099", "#7CCBA2")) + 
    theme_classic() + ylab("Hours") + 
    ggtitle(title) +
    geom_abline(slope = 0, intercept = 0) + 
    theme(
      legend.position = "top",
      axis.title.x = element_blank(),axis.text.x = element_text(size = 20),
      axis.title.y = element_text(size = 20), axis.text.y = element_text(size = 20),
      legend.title = element_blank(), legend.text = element_text(size = 20),
      axis.ticks.x=element_blank()
      ) +
    scale_x_datetime(date_breaks = "2 week", date_labels = "%B %d") 
}



# summarize the result 
wc_result_summary <- function(orig_df, resp_var, wc_sum_df){
  
  scale_pad_summary <- list()
  
  scale_pad_summary[["GPP"]] <- scale_pad(orig_df, resp_var, "GPP")
  scale_pad_summary[["PAR"]] <- scale_pad(orig_df, resp_var, "PPFD_IN_1_1_1")
  scale_pad_summary[["Ts"]] <- scale_pad(orig_df, resp_var, "TS_1_1_1")
  scale_pad_summary[["VWC"]] <- scale_pad(orig_df, resp_var, "SWC_1_1_1")
  
  results <- list()
  
  results[["GPP"]] <- coh_result(scale_pad_summary[["GPP"]], wc_sum_df$GPP)
  results[["PAR"]] <- coh_result(scale_pad_summary[["PAR"]], wc_sum_df$PAR)
  results[["Ts"]] <- coh_result(scale_pad_summary[["Ts"]], wc_sum_df$Ts)
  results[["VWC"]] <- coh_result(scale_pad_summary[["VWC"]], wc_sum_df$VWC)
  
  return(results)
}
```

```{r temporal pwr df & ggplot}
wt_avg <- function(df, ymax){
    
    if (class(df) != "analyze.wavelet") {
        stop("Your object class is not 'analyze.wavelet'")
    }
    
    # Create the dataframe
    output_df <- 
        data.frame(
            Period = df[["Period"]], 
            Power.avg = df[["Power.avg"]],
            Power.avg.pval = df[["Power.avg.pval"]]
        ) %>%
        mutate(Power.avg.sig = ifelse(Power.avg.pval < 0.05, Power.avg, NA))
    
    # Create the plot
    p1 <- 
        ggplot(output_df) + 
        geom_line(aes(log2(Period), Power.avg)) +
        geom_point(aes(log2(Period), Power.avg.sig), size = 3) + 
        theme_classic() + 
        ylab("Avg. Wavelet Spectram") + 
        xlab("Period (days; log scale)") +
        scale_x_continuous(
            breaks = c(0, 3, 6),
            label = c(1, 8, 64)
        ) +
        scale_y_continuous(
            breaks = c(0, 10, 20),
            limits = c(0, ymax)
        ) +
        theme(
            axis.title.x = element_text(size = 25), 
            axis.text.x = element_text(size = 25),
            axis.title.y = element_text(size = 25), 
            axis.text.y = element_text(size = 25),
            legend.title = element_text(size = 25), 
            legend.text = element_text(size = 25),
            panel.border = element_rect(colour = "black", fill = NA, linewidth = 1)
        )
    
    # Return both the dataframe and plot as a list
    return(list(df = output_df, plot = p1))
}


wc_avg <- function(df, ymax){
  
    if (class(df) == "analyze.wavelet") {
      stop("Your object class is 'analyze.wavelet' - use wt_img_gg")
    }
  
   # Create the dataframe
    output_df <- 
        data.frame(
            Period = df[["Period"]], 
            Power.avg = df[["Power.xy.avg"]],
            Power.avg.pval = df[["Power.xy.avg.pval"]]
        ) %>%
        mutate(Power.avg.sig = ifelse(Power.avg.pval < 0.05, Power.avg, NA))
      
      
      p1 <- 
        ggplot(output_df) + 
        geom_line(aes(log2(Period), Power.avg)) +
        geom_point(aes(log2(Period), Power.avg.sig), size = 3) + 
        theme_classic() + 
        ylab("Avg. Wavelet Spectram") + xlab("Period (days; log scale)") +
        scale_x_continuous(
          breaks = c(0, 3, 6),
          label = c(1, 8, 64)
          ) +
        scale_y_continuous(
          breaks = c(0, 10, 20),
          limits = c(0, ymax)
          ) + 
        theme(
          axis.title.x = element_text(size = 25), 
          axis.text.x = element_text(size = 25),
          axis.title.y = element_text(size = 25), 
          axis.text.y = element_text(size = 25),
          legend.title = element_text(size = 25), 
          legend.text = element_text(size = 25),
          panel.border = element_rect(colour = "black", fill = NA, linewidth = 1)
          ) 

    # Return both the dataframe and plot as a list
    return(list(df = output_df, plot = p1))
  
}


```

```{r location to save}
dir_pwr = "G:/My Drive/Research/Projects/CRK_auto/Data/Result/GPP_SR_cospectral/raw_64d_one_GPP_sim100/df_temporal_pwr/"
dir_diurnal = "G:/My Drive/Research/Projects/CRK_auto/Data/Result/GPP_SR_cospectral/raw_64d_one_GPP_sim100/df_diurnal/"
```

```{r CWT MJ2022}
apply(GPP_SR_MJ22, 2, pMiss)

# gap length 
# only PAR
detect_gaps(GPP_SR_MJ22, "GPP")

# Rh
wc_Rh_MJ22 <- wc_summary(GPP_SR_MJ22, "Rh")
wc_Rh_Ts2_MJ22 <- coh(GPP_SR_MJ22, "Rh", "TS_1_2_1", 64)
wc_Rh_VWC2_MJ22 <- coh(GPP_SR_MJ22, "Rh", "SWC_1_2_1", 64)

# heatmap
wc_img_gen(wc_Rh_MJ22, "Rh")

# temporal power data & graph
Rh_GPP_MJ22_pwr <- wc_avg(wc_Rh_MJ22$GPP, 25)
Rh_PAR_MJ22_pwr <- wc_avg(wc_Rh_MJ22$PAR, 25)
Rh_Ts_MJ22_pwr <- wc_avg(wc_Rh_MJ22$Ts, 25)
Rh_VWC_MJ22_pwr <- wc_avg(wc_Rh_MJ22$VWC, 25)

# pwr
Rh_Ts2_MJ22_pwr <- wc_avg(wc_Rh_Ts2_MJ22, 25)
Rh_VWC2_MJ22_pwr <- wc_avg(wc_Rh_VWC2_MJ22, 25)

# csv output
write.csv(Rh_GPP_MJ22_pwr$df, paste0(dir_pwr, "Rh_GPP_MJ22.csv"), row.names = F)
write.csv(Rh_PAR_MJ22_pwr$df, paste0(dir_pwr, "Rh_PAR_MJ22.csv"), row.names = F)
write.csv(Rh_Ts_MJ22_pwr$df, paste0(dir_pwr, "Rh_Ts_MJ22.csv"), row.names = F)
write.csv(Rh_VWC_MJ22_pwr$df, paste0(dir_pwr, "Rh_VWC_MJ22.csv"), row.names = F)

write.csv(Rh_Ts2_MJ22_pwr$df, paste0(dir_pwr, "Rh_Ts20_MJ22.csv"), row.names = F)
write.csv(Rh_VWC2_MJ22_pwr$df, paste0(dir_pwr, "Rh_VWC20_MJ22.csv"), row.names = F)

# extracting diurnal angle & power patterns from heatmap
wc_Rh_MJ22_diurnal_summary <- wc_result_summary(GPP_SR_MJ22, "Rh", wc_Rh_MJ22)
wc_Rh_Ts20_MJ22_diurnal <- coh_result(scale_pad(GPP_SR_MJ22, "Rh", "TS_1_2_1"), wc_Rh_Ts2_MJ22)
wc_Rh_VWC20_MJ22_diurnal <- coh_result(scale_pad(GPP_SR_MJ22, "Rh", "SWC_1_2_1"), wc_Rh_VWC2_MJ22)
# save as csv
save_to_csv(wc_Rh_MJ22_diurnal_summary, dir_diurnal, "wc_Rh_MJ22_")
write.csv(wc_Rh_Ts20_MJ22_diurnal, paste0(dir_diurnal,"wc_Rh_MJ2_Ts20.csv"), row.names = F)
write.csv(wc_Rh_VWC20_MJ22_diurnal, paste0(dir_diurnal,"wc_Rh_MJ22_VWC20.csv"), row.names = F)

# viz
ggarrange(
  wc_lag_img(wc_Rh_MJ22_diurnal_summary$GPP, "GPP"),
  wc_lag_img(wc_Rh_MJ22_diurnal_summary$PAR, "PAR"),
  wc_lag_img(wc_Rh_MJ22_diurnal_summary$Ts, "Ts"),
  wc_lag_img(wc_Rh_MJ22_diurnal_summary$VWC, "VWC"),
  ncol = 2, nrow = 2, common.legend = T
)


#  Ra
wc_Ra_MJ22 <- wc_summary(GPP_SR_MJ22, "Ra1")
wc_Ra_Ts2_MJ22 <- coh(GPP_SR_MJ22, "Ra1", "TS_1_2_1", 64)
wc_Ra_VWC2_MJ22 <- coh(GPP_SR_MJ22, "Ra1", "SWC_1_2_1", 64)

wc_img_gen(wc_Ra_MJ22, "Ra")

# power avg
Ra_GPP_MJ22_pwr <- wc_avg(wc_Ra_MJ22$GPP, 25)
Ra_PAR_MJ22_pwr <- wc_avg(wc_Ra_MJ22$PAR, 25)
Ra_Ts_MJ22_pwr <- wc_avg(wc_Ra_MJ22$Ts, 25)
Ra_VWC_MJ22_pwr <- wc_avg(wc_Ra_MJ22$VWC, 25)

# pwr
Ra_Ts2_MJ22_pwr <- wc_avg(wc_Ra_Ts2_MJ22, 25)
Ra_VWC2_MJ22_pwr <- wc_avg(wc_Ra_VWC2_MJ22, 25)

# csv output
write.csv(Ra_GPP_MJ22_pwr$df, paste0(dir_pwr, "Ra_GPP_MJ22.csv"), row.names = F)
write.csv(Ra_PAR_MJ22_pwr$df, paste0(dir_pwr, "Ra_PAR_MJ22.csv"), row.names = F)
write.csv(Ra_Ts_MJ22_pwr$df, paste0(dir_pwr, "Ra_Ts_MJ22.csv"), row.names = F)
write.csv(Ra_VWC_MJ22_pwr$df, paste0(dir_pwr, "Ra_VWC_MJ22.csv"), row.names = F)


write.csv(Ra_Ts2_MJ22_pwr$df, paste0(dir_pwr, "Ra_Ts20_MJ22.csv"), row.names = F)
write.csv(Ra_VWC2_MJ22_pwr$df, paste0(dir_pwr, "Ra_VWC20_MJ22.csv"), row.names = F)

# extracting diurnal angle & power patterns from heatmap
wc_Ra_MJ22_diurnal_summary <- wc_result_summary(GPP_SR_MJ22, "Ra1", wc_Ra_MJ22)

wc_Ra_Ts20_MJ22_diurnal <- coh_result(scale_pad(GPP_SR_MJ22, "Ra1", "TS_1_2_1"), wc_Rh_Ts2_MJ22)
wc_Ra_VWC20_MJ22_diurnal <- coh_result(scale_pad(GPP_SR_MJ22, "Ra1", "SWC_1_2_1"), wc_Rh_VWC2_MJ22)
# save as csv
save_to_csv(wc_Ra_MJ22_diurnal_summary, dir_diurnal, "wc_Ra_MJ22_")
write.csv(wc_Ra_Ts20_MJ22_diurnal, paste0(dir_diurnal,"wc_Ra_MJ2_Ts20.csv"), row.names = F)
write.csv(wc_Ra_VWC20_MJ22_diurnal, paste0(dir_diurnal,"wc_Ra_MJ22_VWC20.csv"), row.names = F)

# viz
ggarrange(
  wc_lag_img(wc_Ra_MJ22_diurnal_summary$GPP, "GPP"),
  wc_lag_img(wc_Ra_MJ22_diurnal_summary$PAR, "PAR"),
  wc_lag_img(wc_Ra_MJ22_diurnal_summary$Ts, "Ts"),
  wc_lag_img(wc_Ra_MJ22_diurnal_summary$VWC, "VWC"),
  ncol = 2, nrow = 2, common.legend = T
)

```

```{r WT MJ2022}
wt_Rh_MJ22 <- 
  analyze.wavelet(wc_Rh_MJ22$PAR$series,"Rh", 
                  loess.span = 0,     ## no detrending required
                  dt = 1/24,          ## defines the time unit --- there are 24 slots in a day! Time unit = 1 day
                  dj = 1/100,         ## resolution along period axis
                  lowerPeriod = 1/4,  ## 6 hour
                  upperPeriod = 64,  ## 128 days
                  make.pval = T,      ## check for significance
                  n.sim = 100)         ## 10 white-noise simulations used for assessing significance )

Rh_MJ22_pwr <- wt_avg(wt_Rh_MJ22, 25)
write.csv(Rh_MJ22_pwr$df, paste0(dir_pwr, "Rh_MJ22.csv"), row.names = F)

wt.image(wt_Rh_MJ22,  
         legend.params = 
             list(
                 lab = "wavelet power levels", # axis label
                 lab.line = 1.5, # distance between bar label and axis label
                 mar = 5
             ), 
         lvl = 0.3,
         timelab = "",
         show.date = T, 
         date.format = "%Y-%m-%d %H:%M:%S",
         periodlab = "period (days)",
         main = "Rh"
)

# Ra
wt_Ra_MJ22 <- 
  analyze.wavelet(wc_Ra_MJ22$PAR$series,"Ra1", 
                  loess.span = 0,     ## no detrending required
                  dt = 1/24,          ## defines the time unit --- there are 24 slots in a day! Time unit = 1 day
                  dj = 1/100,         ## resolution along period axis
                  lowerPeriod = 1/4,  ## 6 hour
                  upperPeriod = 64,  ## 128 days
                  make.pval = T,      ## check for significance
                  n.sim = 100)         ## 10 white-noise simulations used for assessing significance )

Ra_MJ22_pwr <- wt_avg(wt_Ra_MJ22, 25)
write.csv(Ra_MJ22_pwr$df, paste0(dir_pwr, "Ra_MJ22.csv"), row.names = F)

wt.image(wt_Ra_MJ22,  
         legend.params = 
             list(
                 lab = "wavelet power levels", # axis label
                 lab.line = 1.5, # distance between bar label and axis label
                 mar = 5
             ), 
         lvl = 0.3,
         timelab = "",
         show.date = T, 
         date.format = "%Y-%m-%d %H:%M:%S",
         periodlab = "period (days)",
         main = "Ra"
)
```

```{r CWT MA2023}
apply(GPP_SR_MA23, 2, pMiss)

# Rh
wc_Rh_MA23 <- wc_summary(GPP_SR_MA23, "Rh")
wc_Rh_Ts2_MA23 <- coh(GPP_SR_MA23, "Rh", "TS_1_2_1", 64)
wc_Rh_VWC2_MA23 <- coh(GPP_SR_MA23, "Rh", "SWC_1_2_1", 64)

# heatmap
wc_img_gen(wc_Rh_MA23, "Rh")

# temporal power data & graph
Rh_GPP_MA23_pwr <- wc_avg(wc_Rh_MA23$GPP, 25)
Rh_PAR_MA23_pwr <- wc_avg(wc_Rh_MA23$PAR, 25)
Rh_Ts_MA23_pwr <- wc_avg(wc_Rh_MA23$Ts, 25)
Rh_VWC_MA23_pwr <- wc_avg(wc_Rh_MA23$VWC, 25)

Rh_Ts2_MA23_pwr <- wc_avg(wc_Rh_Ts2_MA23, 25)
Rh_VWC2_MA23_pwr <- wc_avg(wc_Rh_VWC2_MA23, 25)

# csv output
write.csv(Rh_GPP_MA23_pwr$df, paste0(dir_pwr, "Rh_GPP_MA23.csv"), row.names = F)
write.csv(Rh_PAR_MA23_pwr$df, paste0(dir_pwr, "Rh_PAR_MA23.csv"), row.names = F)
write.csv(Rh_Ts_MA23_pwr$df, paste0(dir_pwr, "Rh_Ts_MA23.csv"), row.names = F)
write.csv(Rh_VWC_MA23_pwr$df, paste0(dir_pwr, "Rh_VWC_MA23.csv"), row.names = F)
write.csv(Rh_Ts2_MA23_pwr$df, paste0(dir_pwr, "Rh_Ts20_MA23.csv"), row.names = F)
write.csv(Rh_VWC2_MA23_pwr$df, paste0(dir_pwr, "Rh_VWC20_MA23.csv"), row.names = F)

# extracting diurnal angle & power patterns from heatmap
wc_Rh_MA23_diurnal_summary <- wc_result_summary(GPP_SR_MA23, "Rh", wc_Rh_MA23)
wc_Rh_Ts20_MA23_diurnal <- coh_result(scale_pad(GPP_SR_MA23, "Rh", "TS_1_2_1"), wc_Rh_Ts2_MA23)
wc_Rh_VWC20_MA23_diurnal <- coh_result(scale_pad(GPP_SR_MA23, "Rh", "SWC_1_2_1"), wc_Rh_VWC2_MA23)

# save as csv
save_to_csv(wc_Rh_MA23_diurnal_summary, dir_diurnal, "wc_Rh_MA23_")
write.csv(wc_Rh_Ts20_MA23_diurnal, paste0(dir_diurnal,"wc_Rh_Mar23_Ts20.csv"), row.names = F)
write.csv(wc_Rh_VWC20_MA23_diurnal, paste0(dir_diurnal,"wc_Rh_Mar23_VWC20.csv"), row.names = F)

# viz
ggarrange(
  wc_lag_img(wc_Rh_MA23_diurnal_summary$GPP, "GPP"),
  wc_lag_img(wc_Rh_MA23_diurnal_summary$PAR, "PAR"),
  wc_lag_img(wc_Rh_MA23_diurnal_summary$Ts, "Ts"),
  wc_lag_img(wc_Rh_MA23_diurnal_summary$VWC, "VWC"),
  ncol = 2, nrow = 2, common.legend = T
)


#  Ra
wc_Ra_MA23 <- wc_summary(GPP_SR_MA23, "Ra1")
wc_Ra_Ts2_MA23 <- coh(GPP_SR_MA23, "Ra1", "TS_1_2_1", 64)
wc_Ra_VWC2_MA23 <- coh(GPP_SR_MA23, "Ra1", "SWC_1_2_1", 64)


wc_img_gen(wc_Ra_MA23, "Ra")

# power avg
Ra_GPP_MA23_pwr <- wc_avg(wc_Ra_MA23$GPP, 25)
Ra_PAR_MA23_pwr <- wc_avg(wc_Ra_MA23$PAR, 25)
Ra_Ts_MA23_pwr <- wc_avg(wc_Ra_MA23$Ts, 25)
Ra_VWC_MA23_pwr <- wc_avg(wc_Ra_MA23$VWC, 25)
Ra_Ts2_MA23_pwr <- wc_avg(wc_Ra_Ts2_MA23, 25)
Ra_VWC2_MA23_pwr <- wc_avg(wc_Ra_VWC2_MA23, 25)

# csv output
write.csv(Ra_GPP_MA23_pwr$df, paste0(dir_pwr, "Ra_GPP_MA23.csv"), row.names = F)
write.csv(Ra_PAR_MA23_pwr$df, paste0(dir_pwr, "Ra_PAR_MA23.csv"), row.names = F)
write.csv(Ra_Ts_MA23_pwr$df, paste0(dir_pwr, "Ra_Ts_MA23.csv"), row.names = F)
write.csv(Ra_VWC_MA23_pwr$df, paste0(dir_pwr, "Ra_VWC_MA23.csv"), row.names = F)
write.csv(Ra_Ts2_MA23_pwr$df, paste0(dir_pwr, "Ra_Ts20_MA23.csv"), row.names = F)
write.csv(Ra_VWC2_MA23_pwr$df, paste0(dir_pwr, "Ra_VWC20_MA23.csv"), row.names = F)

# extracting diurnal angle & power patterns from heatmap
wc_Ra_MA23_diurnal_summary <- wc_result_summary(GPP_SR_MA23, "Ra1", wc_Ra_MA23)
wc_Ra_Ts20_MA23_diurnal <- coh_result(scale_pad(GPP_SR_MA23, "Ra1", "TS_1_2_1"), wc_Ra_Ts2_MA23)
wc_Ra_VWC20_MA23_diurnal <- coh_result(scale_pad(GPP_SR_MA23, "Ra1", "SWC_1_2_1"), wc_Ra_VWC2_MA23)

# save as csv
save_to_csv(wc_Ra_MA23_diurnal_summary, dir_diurnal, "wc_Ra_MA23_")

write.csv(wc_Ra_Ts20_MA23_diurnal, paste0(dir_diurnal,"wc_Ra_Mar23_Ts20.csv"), row.names = F)
write.csv(wc_Ra_VWC20_MA23_diurnal, paste0(dir_diurnal,"wc_Ra_Mar23_VWC20.csv"), row.names = F)

# viz
ggarrange(
  wc_lag_img(wc_Ra_MA23_diurnal_summary$GPP, "GPP"),
  wc_lag_img(wc_Ra_MA23_diurnal_summary$PAR, "PAR"),
  wc_lag_img(wc_Ra_MA23_diurnal_summary$Ts, "Ts"),
  wc_lag_img(wc_Ra_MA23_diurnal_summary$VWC, "VWC"),
  ncol = 2, nrow = 2, common.legend = T
)

```

```{r WT MA2023}
wt_Rh_MA23 <- 
  analyze.wavelet(wc_Rh_MA23$PAR$series,"Rh", 
                  loess.span = 0,     ## no detrending required
                  dt = 1/24,          ## defines the time unit --- there are 24 slots in a day! Time unit = 1 day
                  dj = 1/100,         ## resolution along period axis
                  lowerPeriod = 1/4,  ## 6 hour
                  upperPeriod = 64,  ## 128 days
                  make.pval = T,      ## check for significance
                  n.sim = 100)         ## 10 white-noise simulations used for assessing significance )

Rh_MA23_pwr <- wt_avg(wt_Rh_MA23, 25)
write.csv(Rh_MA23_pwr$df, paste0(dir_pwr, "Rh_MA23.csv"), row.names = F)

wt.image(wt_Rh_MA23,  
         legend.params = 
             list(
                 lab = "wavelet power levels", # axis label
                 lab.line = 1.5, # distance between bar label and axis label
                 mar = 5
             ), 
         lvl = 0.3,
         timelab = "",
         show.date = T, 
         date.format = "%Y-%m-%d %H:%M:%S",
         periodlab = "period (days)",
         main = "Rh"
)

# Ra
wt_Ra_MA23 <- 
  analyze.wavelet(wc_Ra_MA23$PAR$series,"Ra1", 
                  loess.span = 0,     ## no detrending required
                  dt = 1/24,          ## defines the time unit --- there are 24 slots in a day! Time unit = 1 day
                  dj = 1/100,         ## resolution along period axis
                  lowerPeriod = 1/4,  ## 6 hour
                  upperPeriod = 64,  ## 128 days
                  make.pval = T,      ## check for significance
                  n.sim = 10)         ## 10 white-noise simulations used for assessing significance )

Ra_MA23_pwr <- wt_avg(wt_Ra_MA23, 25)
write.csv(Ra_MA23_pwr$df, paste0(dir_pwr, "Ra_MA23.csv"), row.names = F)

wt.image(wt_Ra_MA23,  
         legend.params = 
             list(
                 lab = "wavelet power levels", # axis label
                 lab.line = 1.5, # distance between bar label and axis label
                 mar = 5
             ), 
         lvl = 0.3,
         timelab = "",
         show.date = T, 
         date.format = "%Y-%m-%d %H:%M:%S",
         periodlab = "period (days)",
         main = "Ra"
)
```

```{r CWT Mar2024}
apply(GPP_SR_Mar24, 2, pMiss)

# Rh
wc_Rh_Mar24 <- wc_summary(GPP_SR_Mar24, "Rh")
wc_Rh_Ts2_Mar24 <- coh(GPP_SR_Mar24, "Rh", "TS_1_2_1", 64)
wc_Rh_VWC2_Mar24 <- coh(GPP_SR_Mar24, "Rh", "SWC_1_2_1", 64)

# heatmap
wc_img_gen(wc_Rh_Mar24, "Rh")

# temporal power data & graph
Rh_GPP_Mar24_pwr <- wc_avg(wc_Rh_Mar24$GPP, 25)
Rh_PAR_Mar24_pwr <- wc_avg(wc_Rh_Mar24$PAR, 25)
Rh_Ts_Mar24_pwr <- wc_avg(wc_Rh_Mar24$Ts, 25)
Rh_VWC_Mar24_pwr <- wc_avg(wc_Rh_Mar24$VWC, 25)
Rh_Ts2_Mar24_pwr <- wc_avg(wc_Rh_Ts2_Mar24, 25)
Rh_VWC2_Mar24_pwr <- wc_avg(wc_Rh_VWC2_Mar24, 25)

# csv output
write.csv(Rh_GPP_Mar24_pwr$df, paste0(dir_pwr, "Rh_GPP_Mar24.csv"), row.names = F)
write.csv(Rh_PAR_Mar24_pwr$df, paste0(dir_pwr, "Rh_PAR_Mar24.csv"), row.names = F)
write.csv(Rh_Ts_Mar24_pwr$df, paste0(dir_pwr, "Rh_Ts_Mar24.csv"), row.names = F)
write.csv(Rh_VWC_Mar24_pwr$df, paste0(dir_pwr, "Rh_VWC_Mar24.csv"), row.names = F)
write.csv(Rh_Ts2_Mar24_pwr$df, paste0(dir_pwr, "Rh_Ts20_Mar24.csv"), row.names = F)
write.csv(Rh_VWC2_Mar24_pwr$df, paste0(dir_pwr, "Rh_VWC20_Mar24.csv"), row.names = F)

# extracting diurnal angle & power patterns from heatmap
wc_Rh_Mar24_diurnal_summary <- wc_result_summary(GPP_SR_Mar24, "Rh", wc_Rh_Mar24)
wc_Rh_Ts20_Mar24_diurnal <- coh_result(scale_pad(GPP_SR_Mar24, "Rh", "TS_1_2_1"), wc_Rh_Ts2_Mar24)
wc_Rh_VWC20_Mar24_diurnal <- coh_result(scale_pad(GPP_SR_Mar24, "Rh", "SWC_1_2_1"), wc_Rh_VWC2_Mar24)
# save as csv
save_to_csv(wc_Rh_Mar24_diurnal_summary, dir_diurnal, "wc_Rh_Mar24_")
write.csv(wc_Rh_Ts20_Mar24_diurnal, paste0(dir_diurnal,"wc_Rh_Mar24_Ts20.csv"), row.names = F)
write.csv(wc_Rh_VWC20_Mar24_diurnal, paste0(dir_diurnal,"wc_Rh_Mar24_VWC20.csv"), row.names = F)

# viz
ggarrange(
  wc_lag_img(wc_Rh_Mar24_diurnal_summary$GPP, "GPP"),
  wc_lag_img(wc_Rh_Mar24_diurnal_summary$PAR, "PAR"),
  wc_lag_img(wc_Rh_Mar24_diurnal_summary$Ts, "Ts"),
  wc_lag_img(wc_Rh_Mar24_diurnal_summary$VWC, "VWC"),
  ncol = 2, nrow = 2, common.legend = T
)


#  Ra
wc_Ra_Mar24 <- wc_summary(GPP_SR_Mar24, "Ra")
wc_Ra_Ts2_Mar24 <- coh(GPP_SR_Mar24, "Ra", "TS_1_2_1", 64)
wc_Ra_VWC2_Mar24 <- coh(GPP_SR_Mar24, "Ra", "SWC_1_2_1", 64)

wc_img_gen(wc_Ra_Mar24, "Ra")

# power avg
Ra_GPP_Mar24_pwr <- wc_avg(wc_Ra_Mar24$GPP, 25)
Ra_PAR_Mar24_pwr <- wc_avg(wc_Ra_Mar24$PAR, 25)
Ra_Ts_Mar24_pwr <- wc_avg(wc_Ra_Mar24$Ts, 25)
Ra_VWC_Mar24_pwr <- wc_avg(wc_Ra_Mar24$VWC, 25)
Ra_Ts2_Mar24_pwr <- wc_avg(wc_Ra_Ts2_Mar24, 25)
Ra_VWC2_Mar24_pwr <- wc_avg(wc_Ra_VWC2_Mar24, 25)


# csv output
write.csv(Ra_GPP_Mar24_pwr$df, paste0(dir_pwr, "Ra_GPP_Mar24.csv"), row.names = F)
write.csv(Ra_PAR_Mar24_pwr$df, paste0(dir_pwr, "Ra_PAR_Mar24.csv"), row.names = F)
write.csv(Ra_Ts_Mar24_pwr$df, paste0(dir_pwr, "Ra_Ts_Mar24.csv"), row.names = F)
write.csv(Ra_VWC_Mar24_pwr$df, paste0(dir_pwr, "Ra_VWC_Mar24.csv"), row.names = F)
write.csv(Ra_Ts2_Mar24_pwr$df, paste0(dir_pwr, "Ra_Ts20_Mar24.csv"), row.names = F)
write.csv(Ra_VWC2_Mar24_pwr$df, paste0(dir_pwr, "Ra_VWC20_Mar24.csv"), row.names = F)

# extracting diurnal angle & power patterns from heatmap
wc_Ra_Mar24_diurnal_summary <- wc_result_summary(GPP_SR_Mar24, "Ra", wc_Ra_Mar24)
wc_Ra_Ts20_Mar24_diurnal <- coh_result(scale_pad(GPP_SR_Mar24, "Ra", "TS_1_2_1"), wc_Ra_Ts2_Mar24)
wc_Ra_VWC20_Mar24_diurnal <- coh_result(scale_pad(GPP_SR_Mar24, "Ra", "SWC_1_2_1"), wc_Ra_VWC2_Mar24)
# save as csv
save_to_csv(wc_Ra_Mar24_diurnal_summary, dir_diurnal, "wc_Ra_Mar24_")
write.csv(wc_Ra_Ts20_Mar24_diurnal, paste0(dir_diurnal,"wc_Ra_Mar24_Ts20.csv"), row.names = F)
write.csv(wc_Ra_VWC20_Mar24_diurnal, paste0(dir_diurnal,"wc_Ra_Mar24_VWC20.csv"), row.names = F)

# viz
ggarrange(
  wc_lag_img(wc_Ra_Mar24_diurnal_summary$GPP, "GPP"),
  wc_lag_img(wc_Ra_Mar24_diurnal_summary$PAR, "PAR"),
  wc_lag_img(wc_Ra_Mar24_diurnal_summary$Ts, "Ts"),
  wc_lag_img(wc_Ra_Mar24_diurnal_summary$VWC, "VWC"),
  ncol = 2, nrow = 2, common.legend = T
)

```

```{r WT Mar2024}
wt_Rh_Mar24 <- 
  analyze.wavelet(wc_Rh_Mar24$PAR$series,"Rh", 
                  loess.span = 0,     ## no detrending required
                  dt = 1/24,          ## defines the time unit --- there are 24 slots in a day! Time unit = 1 day
                  dj = 1/100,         ## resolution along period axis
                  lowerPeriod = 1/4,  ## 6 hour
                  upperPeriod = 64,  ## 128 days
                  make.pval = T,      ## check for significance
                  n.sim = 100)         ## 10 white-noise simulations used for assessing significance )

Rh_Mar24_pwr <- wt_avg(wt_Rh_Mar24, 25)
write.csv(Rh_Mar24_pwr$df, paste0(dir_pwr, "Rh_Mar24.csv"), row.names = F)

wt.image(wt_Rh_Mar24,  
         legend.params = 
             list(
                 lab = "wavelet power levels", # axis label
                 lab.line = 1.5, # distance between bar label and axis label
                 mar = 5
             ), 
         lvl = 0.3,
         timelab = "",
         show.date = T, 
         date.format = "%Y-%m-%d %H:%M:%S",
         periodlab = "period (days)",
         main = "Rh"
)

# Ra
wt_Ra_Mar24 <- 
  analyze.wavelet(wc_Ra_Mar24$PAR$series,"Ra", 
                  loess.span = 0,     ## no detrending required
                  dt = 1/24,          ## defines the time unit --- there are 24 slots in a day! Time unit = 1 day
                  dj = 1/100,         ## resolution along period axis
                  lowerPeriod = 1/4,  ## 6 hour
                  upperPeriod = 64,  ## 128 days
                  make.pval = T,      ## check for significance
                  n.sim = 100)         ## 10 white-noise simulations used for assessing significance )

Ra_Mar24_pwr <- wt_avg(wt_Ra_Mar24, 25)
write.csv(Ra_Mar24_pwr$df, paste0(dir_pwr, "Ra_Mar24.csv"), row.names = F)

wt.image(wt_Ra_Mar24,  
         legend.params = 
             list(
                 lab = "wavelet power levels", # axis label
                 lab.line = 1.5, # distance between bar label and axis label
                 mar = 5
             ), 
         lvl = 0.3,
         timelab = "",
         show.date = T, 
         date.format = "%Y-%m-%d %H:%M:%S",
         periodlab = "period (days)",
         main = "Ra"
)
```

```{r CWT Apr2024}
apply(GPP_SR_Apr24, 2, pMiss)

# gap length 
# only PAR
detect_gaps(GPP_SR_Apr24, "PPFD_IN_1_1_1")
detect_gaps(GPP_SR_Apr24, "GPP")

# Rh
wc_Rh_Apr24 <- wc_summary(GPP_SR_Apr24, "Rh")
wc_Rh_Ts2_Apr24 <- coh(GPP_SR_Apr24, "Rh", "TS_1_2_1", 64)
wc_Rh_VWC2_Apr24 <- coh(GPP_SR_Apr24, "Rh", "SWC_1_2_1", 64)

# heatmap
wc_img_gen(wc_Rh_Apr24, "Rh")

# temporal power data & graph
Rh_GPP_Apr24_pwr <- wc_avg(wc_Rh_Apr24$GPP, 25)
Rh_PAR_Apr24_pwr <- wc_avg(wc_Rh_Apr24$PAR, 25)
Rh_Ts_Apr24_pwr <- wc_avg(wc_Rh_Apr24$Ts, 25)
Rh_VWC_Apr24_pwr <- wc_avg(wc_Rh_Apr24$VWC, 25)
Rh_Ts2_Apr24_pwr <- wc_avg(wc_Rh_Ts2_Apr24, 25)
Rh_VWC2_Apr24_pwr <- wc_avg(wc_Rh_VWC2_Apr24, 25)

# csv output
write.csv(Rh_GPP_Apr24_pwr$df, paste0(dir_pwr, "Rh_GPP_Apr24.csv"), row.names = F)
write.csv(Rh_PAR_Apr24_pwr$df, paste0(dir_pwr, "Rh_PAR_Apr24.csv"), row.names = F)
write.csv(Rh_Ts_Apr24_pwr$df, paste0(dir_pwr, "Rh_Ts_Apr24.csv"), row.names = F)
write.csv(Rh_VWC_Apr24_pwr$df, paste0(dir_pwr, "Rh_VWC_Apr24.csv"), row.names = F)
write.csv(Rh_Ts2_Apr24_pwr$df, paste0(dir_pwr, "Rh_Ts20_Apr24.csv"), row.names = F)
write.csv(Rh_VWC2_Apr24_pwr$df, paste0(dir_pwr, "Rh_VWC20_Apr24.csv"), row.names = F)

# extracting diurnal angle & power patterns from heatmap
wc_Rh_Apr24_diurnal_summary <- wc_result_summary(GPP_SR_Apr24, "Rh", wc_Rh_Apr24)
wc_Rh_Ts20_Apr24_diurnal <- coh_result(scale_pad(GPP_SR_Apr24, "Rh", "TS_1_2_1"), wc_Rh_Ts2_Apr24)
wc_Rh_VWC20_Apr24_diurnal <- coh_result(scale_pad(GPP_SR_Apr24, "Rh", "SWC_1_2_1"), wc_Rh_VWC2_Apr24)

# save as csv
save_to_csv(wc_Rh_Apr24_diurnal_summary, dir_diurnal, "wc_Rh_Apr24_")
write.csv(wc_Rh_Ts20_Apr24_diurnal, paste0(dir_diurnal,"wc_Rh_Apr24_Ts20.csv"), row.names = F)
write.csv(wc_Rh_VWC20_Apr24_diurnal, paste0(dir_diurnal,"wc_Rh_Apr24_VWC20.csv"), row.names = F)

# viz
ggarrange(
  wc_lag_img(wc_Rh_Apr24_diurnal_summary$GPP, "GPP"),
  wc_lag_img(wc_Rh_Apr24_diurnal_summary$PAR, "PAR"),
  wc_lag_img(wc_Rh_Apr24_diurnal_summary$Ts, "Ts"),
  wc_lag_img(wc_Rh_Apr24_diurnal_summary$VWC, "VWC"),
  ncol = 2, nrow = 2, common.legend = T
)


#  Ra
wc_Ra_Apr24 <- wc_summary(GPP_SR_Apr24, "Ra")
wc_Ra_Ts2_Apr24 <- coh(GPP_SR_Apr24, "Ra", "TS_1_2_1", 64)
wc_Ra_VWC2_Apr24 <- coh(GPP_SR_Apr24, "Ra", "SWC_1_2_1", 64)

wc_img_gen(wc_Ra_Apr24, "Ra")

# power avg
Ra_GPP_Apr24_pwr <- wc_avg(wc_Ra_Apr24$GPP, 25)
Ra_PAR_Apr24_pwr <- wc_avg(wc_Ra_Apr24$PAR, 25)
Ra_Ts_Apr24_pwr <- wc_avg(wc_Ra_Apr24$Ts, 25)
Ra_VWC_Apr24_pwr <- wc_avg(wc_Ra_Apr24$VWC, 25)
Ra_Ts2_Apr24_pwr <- wc_avg(wc_Ra_Ts2_Apr24, 25)
Ra_VWC2_Apr24_pwr <- wc_avg(wc_Ra_VWC2_Apr24, 25)

# csv output
write.csv(Ra_GPP_Apr24_pwr$df, paste0(dir_pwr, "Ra_GPP_Apr24.csv"), row.names = F)
write.csv(Ra_PAR_Apr24_pwr$df, paste0(dir_pwr, "Ra_PAR_Apr24.csv"), row.names = F)
write.csv(Ra_Ts_Apr24_pwr$df, paste0(dir_pwr, "Ra_Ts_Apr24.csv"), row.names = F)
write.csv(Ra_VWC_Apr24_pwr$df, paste0(dir_pwr, "Ra_VWC_Apr24.csv"), row.names = F)
write.csv(Ra_Ts2_Apr24_pwr$df, paste0(dir_pwr, "Ra_Ts20_Apr24.csv"), row.names = F)
write.csv(Ra_VWC2_Apr24_pwr$df, paste0(dir_pwr, "Ra_VWC20_Apr24.csv"), row.names = F)

# extracting diurnal angle & power patterns from heatmap
wc_Ra_Apr24_diurnal_summary <- wc_result_summary(GPP_SR_Apr24, "Ra", wc_Ra_Apr24)
wc_Ra_Ts20_Apr24_diurnal <- coh_result(scale_pad(GPP_SR_Apr24, "Ra", "TS_1_2_1"), wc_Ra_Ts2_Apr24)
wc_Ra_VWC20_Apr24_diurnal <- coh_result(scale_pad(GPP_SR_Apr24, "Ra", "SWC_1_2_1"), wc_Ra_VWC2_Apr24)
# save as csv
save_to_csv(wc_Ra_Apr24_diurnal_summary, dir_diurnal, "wc_Ra_Apr24_")

# viz
ggarrange(
  wc_lag_img(wc_Ra_Apr24_diurnal_summary$GPP, "GPP"),
  wc_lag_img(wc_Ra_Apr24_diurnal_summary$PAR, "PAR"),
  wc_lag_img(wc_Ra_Apr24_diurnal_summary$Ts, "Ts"),
  wc_lag_img(wc_Ra_Apr24_diurnal_summary$VWC, "VWC"),
  ncol = 2, nrow = 2, common.legend = T
)

```

```{r WT Apr2024}
wt_Rh_Apr24 <- 
  analyze.wavelet(wc_Rh_Apr24$PAR$series,"Rh", 
                  loess.span = 0,     ## no detrending required
                  dt = 1/24,          ## defines the time unit --- there are 24 slots in a day! Time unit = 1 day
                  dj = 1/100,         ## resolution along period axis
                  lowerPeriod = 1/4,  ## 6 hour
                  upperPeriod = 64,  ## 128 days
                  make.pval = T,      ## check for significance
                  n.sim = 100)         ## 10 white-noise simulations used for assessing significance )

Rh_Apr24_pwr <- wt_avg(wt_Rh_Apr24, 25)
write.csv(Rh_Apr24_pwr$df, paste0(dir_pwr, "Rh_Apr24.csv"), row.names = F)

wt.image(wt_Rh_Apr24,  
         legend.params = 
             list(
                 lab = "wavelet power levels", # axis label
                 lab.line = 1.5, # distance between bar label and axis label
                 mar = 5
             ), 
         lvl = 0.3,
         timelab = "",
         show.date = T, 
         date.format = "%Y-%m-%d %H:%M:%S",
         periodlab = "period (days)",
         main = "Rh"
)

# Ra
wt_Ra_Apr24 <- 
  analyze.wavelet(wc_Ra_Apr24$PAR$series,"Ra", 
                  loess.span = 0,     ## no detrending required
                  dt = 1/24,          ## defines the time unit --- there are 24 slots in a day! Time unit = 1 day
                  dj = 1/100,         ## resolution along period axis
                  lowerPeriod = 1/4,  ## 6 hour
                  upperPeriod = 64,  ## 128 days
                  make.pval = T,      ## check for significance
                  n.sim = 100)         ## 10 white-noise simulations used for assessing significance )

Ra_Apr24_pwr <- wt_avg(wt_Ra_Apr24, 25)
write.csv(Ra_Apr24_pwr$df, paste0(dir_pwr, "Ra_Apr24.csv"), row.names = F)

wt.image(wt_Ra_Apr24,  
         legend.params = 
             list(
                 lab = "wavelet power levels", # axis label
                 lab.line = 1.5, # distance between bar label and axis label
                 mar = 5
             ), 
         lvl = 0.3,
         timelab = "",
         show.date = T, 
         date.format = "%Y-%m-%d %H:%M:%S",
         periodlab = "period (days)",
         main = "Ra"
)
```

```{r CWT Sep2024}
apply(GPP_SR_Sep24, 2, pMiss)

# Rh
wc_Rh_Sep24 <- wc_summary(GPP_SR_Sep24, "Rh")
wc_Rh_Ts2_Sep24 <- coh(GPP_SR_Sep24, "Rh", "TS_1_2_1", 64)
wc_Rh_VWC2_Sep24 <- coh(GPP_SR_Sep24, "Rh", "SWC_1_2_1", 64)

# heatmap
wc_img_gen(wc_Rh_Sep24, "Rh")

# temporal power data & graph
Rh_GPP_Sep24_pwr <- wc_avg(wc_Rh_Sep24$GPP, 25)
Rh_PAR_Sep24_pwr <- wc_avg(wc_Rh_Sep24$PAR, 25)
Rh_Ts_Sep24_pwr <- wc_avg(wc_Rh_Sep24$Ts, 25)
Rh_VWC_Sep24_pwr <- wc_avg(wc_Rh_Sep24$VWC, 25)
Rh_Ts2_Sep24_pwr <- wc_avg(wc_Rh_Ts2_Sep24, 25)
Rh_VWC2_Sep24_pwr <- wc_avg(wc_Rh_VWC2_Sep24, 25)

# csv output
write.csv(Rh_GPP_Sep24_pwr$df, paste0(dir_pwr, "Rh_GPP_Sep24.csv"), row.names = F)
write.csv(Rh_PAR_Sep24_pwr$df, paste0(dir_pwr, "Rh_PAR_Sep24.csv"), row.names = F)
write.csv(Rh_Ts_Sep24_pwr$df, paste0(dir_pwr, "Rh_Ts_Sep24.csv"), row.names = F)
write.csv(Rh_VWC_Sep24_pwr$df, paste0(dir_pwr, "Rh_VWC_Sep24.csv"), row.names = F)
write.csv(Rh_Ts2_Sep24_pwr$df, paste0(dir_pwr, "Rh_Ts20_Sep24.csv"), row.names = F)
write.csv(Rh_VWC2_Sep24_pwr$df, paste0(dir_pwr, "Rh_VWC20_Sep24.csv"), row.names = F)

# extracting diurnal angle & power patterns from heatmap
wc_Rh_Sep24_diurnal_summary <- wc_result_summary(GPP_SR_Sep24, "Rh", wc_Rh_Sep24)
wc_Rh_Ts20_Sep24_diurnal <- coh_result(scale_pad(GPP_SR_Sep24, "Rh", "TS_1_2_1"), wc_Rh_Ts2_Sep24)
wc_Rh_VWC20_Sep24_diurnal <- coh_result(scale_pad(GPP_SR_Sep24, "Rh", "SWC_1_2_1"), wc_Rh_VWC2_Sep24)
# save as csv
save_to_csv(wc_Rh_Sep24_diurnal_summary, dir_diurnal, "wc_Rh_Sep24_")
write.csv(wc_Rh_Ts20_Sep24_diurnal, paste0(dir_diurnal,"wc_Rh_Sep24_Ts20.csv"), row.names = F)
write.csv(wc_Rh_VWC20_Sep24_diurnal, paste0(dir_diurnal,"wc_Rh_Sep24_VWC20.csv"), row.names = F)
# viz
ggarrange(
  wc_lag_img(wc_Rh_Sep24_diurnal_summary$GPP, "GPP"),
  wc_lag_img(wc_Rh_Sep24_diurnal_summary$PAR, "PAR"),
  wc_lag_img(wc_Rh_Sep24_diurnal_summary$Ts, "Ts"),
  wc_lag_img(wc_Rh_Sep24_diurnal_summary$VWC, "VWC"),
  ncol = 2, nrow = 2, common.legend = T
)


#  Ra
wc_Ra_Sep24 <- wc_summary(GPP_SR_Sep24, "Ra")
wc_Ra_Ts2_Sep24 <- coh(GPP_SR_Sep24, "Ra", "TS_1_2_1", 64)
wc_Ra_VWC2_Sep24 <- coh(GPP_SR_Sep24, "Ra", "SWC_1_2_1", 64)

wc_img_gen(wc_Ra_Sep24, "Ra")

# power avg
Ra_GPP_Sep24_pwr <- wc_avg(wc_Ra_Sep24$GPP, 25)
Ra_PAR_Sep24_pwr <- wc_avg(wc_Ra_Sep24$PAR, 25)
Ra_Ts_Sep24_pwr <- wc_avg(wc_Ra_Sep24$Ts, 25)
Ra_VWC_Sep24_pwr <- wc_avg(wc_Ra_Sep24$VWC, 25)
Ra_Ts2_Sep24_pwr <- wc_avg(wc_Ra_Ts2_Sep24, 25)
Ra_VWC2_Sep24_pwr <- wc_avg(wc_Ra_VWC2_Sep24, 25)

# csv output
write.csv(Ra_GPP_Sep24_pwr$df, paste0(dir_pwr, "Ra_GPP_Sep24.csv"), row.names = F)
write.csv(Ra_PAR_Sep24_pwr$df, paste0(dir_pwr, "Ra_PAR_Sep24.csv"), row.names = F)
write.csv(Ra_Ts_Sep24_pwr$df, paste0(dir_pwr, "Ra_Ts_Sep24.csv"), row.names = F)
write.csv(Ra_VWC_Sep24_pwr$df, paste0(dir_pwr, "Ra_VWC_Sep24.csv"), row.names = F)
write.csv(Ra_Ts2_Sep24_pwr$df, paste0(dir_pwr, "Ra_Ts20_Sep24.csv"), row.names = F)
write.csv(Ra_VWC2_Sep24_pwr$df, paste0(dir_pwr, "Ra_VWC20_Sep24.csv"), row.names = F)

# extracting diurnal angle & power patterns from heatmap
wc_Ra_Sep24_diurnal_summary <- wc_result_summary(GPP_SR_Sep24, "Ra", wc_Ra_Sep24)
wc_Ra_Ts20_Sep24_diurnal <- coh_result(scale_pad(GPP_SR_Sep24, "Ra", "TS_1_2_1"), wc_Ra_Ts2_Sep24)
wc_Ra_VWC20_Sep24_diurnal <- coh_result(scale_pad(GPP_SR_Sep24, "Ra", "SWC_1_2_1"), wc_Ra_VWC2_Sep24)

# save as csv
save_to_csv(wc_Ra_Sep24_diurnal_summary, dir_diurnal, "wc_Ra_Sep24_")
write.csv(wc_Ra_Ts20_Sep24_diurnal, paste0(dir_diurnal,"wc_Ra_Sep24_Ts20.csv"), row.names = F)
write.csv(wc_Ra_VWC20_Sep24_diurnal, paste0(dir_diurnal,"wc_Ra_Sep24_VWC20.csv"), row.names = F)

# viz
ggarrange(
  wc_lag_img(wc_Ra_Sep24_diurnal_summary$GPP, "GPP"),
  wc_lag_img(wc_Ra_Sep24_diurnal_summary$PAR, "PAR"),
  wc_lag_img(wc_Ra_Sep24_diurnal_summary$Ts, "Ts"),
  wc_lag_img(wc_Ra_Sep24_diurnal_summary$VWC, "VWC"),
  ncol = 2, nrow = 2, common.legend = T
)

```

```{r WT Sep2024}
wt_Rh_Sep24 <- 
  analyze.wavelet(wc_Rh_Sep24$PAR$series,"Rh", 
                  loess.span = 0,     ## no detrending required
                  dt = 1/24,          ## defines the time unit --- there are 24 slots in a day! Time unit = 1 day
                  dj = 1/100,         ## resolution along period axis
                  lowerPeriod = 1/4,  ## 6 hour
                  upperPeriod = 64,  ## 128 days
                  make.pval = T,      ## check for significance
                  n.sim = 100)         ## 10 white-noise simulations used for assessing significance )

Rh_Sep24_pwr <- wt_avg(wt_Rh_Sep24, 25)
write.csv(Rh_Sep24_pwr$df, paste0(dir_pwr, "Rh_Sep24.csv"), row.names = F)

wt.image(wt_Rh_Sep24,  
         legend.params = 
             list(
                 lab = "wavelet power levels", # axis label
                 lab.line = 1.5, # distance between bar label and axis label
                 mar = 5
             ), 
         lvl = 0.3,
         timelab = "",
         show.date = T, 
         date.format = "%Y-%m-%d %H:%M:%S",
         periodlab = "period (days)",
         main = "Rh"
)

# Ra
wt_Ra_Sep24 <- 
  analyze.wavelet(wc_Ra_Sep24$PAR$series,"Ra", 
                  loess.span = 0,     ## no detrending required
                  dt = 1/24,          ## defines the time unit --- there are 24 slots in a day! Time unit = 1 day
                  dj = 1/100,         ## resolution along period axis
                  lowerPeriod = 1/4,  ## 6 hour
                  upperPeriod = 64,  ## 128 days
                  make.pval = T,      ## check for significance
                  n.sim = 100)         ## 10 white-noise simulations used for assessing significance )

Ra_Sep24_pwr <- wt_avg(wt_Ra_Sep24, 25)
write.csv(Ra_Sep24_pwr$df, paste0(dir_pwr, "Ra_Sep24.csv"), row.names = F)

wt.image(wt_Ra_Sep24,  
         legend.params = 
             list(
                 lab = "wavelet power levels", # axis label
                 lab.line = 1.5, # distance between bar label and axis label
                 mar = 5
             ), 
         lvl = 0.3,
         timelab = "",
         show.date = T, 
         date.format = "%Y-%m-%d %H:%M:%S",
         periodlab = "period (days)",
         main = "Ra"
)
```

```{r CWT Oct2024}
apply(GPP_SR_Oct24, 2, pMiss)

# gap length 
# only PAR
detect_gaps(GPP_SR_Oct24, "GPP")

# Rh
wc_Rh_Oct24 <- wc_summary(GPP_SR_Oct24, "Rh")
wc_Rh_Ts2_Oct24 <- coh(GPP_SR_Oct24, "Rh", "TS_1_2_1", 64)
wc_Rh_VWC2_Oct24 <- coh(GPP_SR_Oct24, "Rh", "SWC_1_2_1", 64)

# heatmap
wc_img_gen(wc_Rh_Oct24, "Rh")

# temporal power data & graph
Rh_GPP_Oct24_pwr <- wc_avg(wc_Rh_Oct24$GPP, 25)
Rh_PAR_Oct24_pwr <- wc_avg(wc_Rh_Oct24$PAR, 25)
Rh_Ts_Oct24_pwr <- wc_avg(wc_Rh_Oct24$Ts, 25)
Rh_VWC_Oct24_pwr <- wc_avg(wc_Rh_Oct24$VWC, 25)
Rh_Ts2_Oct24_pwr <- wc_avg(wc_Rh_Ts2_Oct24, 25)
Rh_VWC2_Oct24_pwr <- wc_avg(wc_Rh_VWC2_Oct24, 25)

# csv output
write.csv(Rh_GPP_Oct24_pwr$df, paste0(dir_pwr, "Rh_GPP_Oct24.csv"), row.names = F)
write.csv(Rh_PAR_Oct24_pwr$df, paste0(dir_pwr, "Rh_PAR_Oct24.csv"), row.names = F)
write.csv(Rh_Ts_Oct24_pwr$df, paste0(dir_pwr, "Rh_Ts_Oct24.csv"), row.names = F)
write.csv(Rh_VWC_Oct24_pwr$df, paste0(dir_pwr, "Rh_VWC_Oct24.csv"), row.names = F)
write.csv(Rh_Ts2_Oct24_pwr$df, paste0(dir_pwr, "Rh_Ts20_Oct24.csv"), row.names = F)
write.csv(Rh_VWC2_Oct24_pwr$df, paste0(dir_pwr, "Rh_VWC20_Oct24.csv"), row.names = F)


# extracting diurnal angle & power patterns from heatmap
wc_Rh_Oct24_diurnal_summary <- wc_result_summary(GPP_SR_Oct24, "Rh", wc_Rh_Oct24)
wc_Rh_Ts20_Oct24_diurnal <- coh_result(scale_pad(GPP_SR_Oct24, "Rh", "TS_1_2_1"), wc_Rh_Ts2_Oct24)
wc_Rh_VWC20_Oct24_diurnal <- coh_result(scale_pad(GPP_SR_Oct24, "Rh", "SWC_1_2_1"), wc_Rh_VWC2_Oct24)
# save as csv
save_to_csv(wc_Rh_Oct24_diurnal_summary, dir_diurnal, "wc_Rh_Oct24_")
write.csv(wc_Rh_Ts20_Oct24_diurnal, paste0(dir_diurnal,"wc_Rh_Oct24_Ts20.csv"), row.names = F)
write.csv(wc_Rh_VWC20_Oct24_diurnal, paste0(dir_diurnal,"wc_Rh_Oct24_VWC20.csv"), row.names = F)

# viz
ggarrange(
  wc_lag_img(wc_Rh_Oct24_diurnal_summary$GPP, "GPP"),
  wc_lag_img(wc_Rh_Oct24_diurnal_summary$PAR, "PAR"),
  wc_lag_img(wc_Rh_Oct24_diurnal_summary$Ts, "Ts"),
  wc_lag_img(wc_Rh_Oct24_diurnal_summary$VWC, "VWC"),
  ncol = 2, nrow = 2, common.legend = T
)


#  Ra
wc_Ra_Oct24 <- wc_summary(GPP_SR_Oct24, "Ra")
wc_Ra_Ts2_Oct24 <- coh(GPP_SR_Oct24, "Ra", "TS_1_2_1", 64)
wc_Ra_VWC2_Oct24 <- coh(GPP_SR_Oct24, "Ra", "SWC_1_2_1", 64)

wc_img_gen(wc_Ra_Oct24, "Ra")

# power avg
Ra_GPP_Oct24_pwr <- wc_avg(wc_Ra_Oct24$GPP, 25)
Ra_PAR_Oct24_pwr <- wc_avg(wc_Ra_Oct24$PAR, 25)
Ra_Ts_Oct24_pwr <- wc_avg(wc_Ra_Oct24$Ts, 25)
Ra_VWC_Oct24_pwr <- wc_avg(wc_Ra_Oct24$VWC, 25)
Ra_Ts2_Oct24_pwr <- wc_avg(wc_Ra_Ts2_Oct24, 25)
Ra_VWC2_Oct24_pwr <- wc_avg(wc_Ra_VWC2_Oct24, 25)

# csv output
write.csv(Ra_GPP_Oct24_pwr$df, paste0(dir_pwr, "Ra_GPP_Oct24.csv"), row.names = F)
write.csv(Ra_PAR_Oct24_pwr$df, paste0(dir_pwr, "Ra_PAR_Oct24.csv"), row.names = F)
write.csv(Ra_Ts_Oct24_pwr$df, paste0(dir_pwr, "Ra_Ts_Oct24.csv"), row.names = F)
write.csv(Ra_VWC_Oct24_pwr$df, paste0(dir_pwr, "Ra_VWC_Oct24.csv"), row.names = F)
write.csv(Ra_Ts2_Oct24_pwr$df, paste0(dir_pwr, "Ra_Ts20_Oct24.csv"), row.names = F)
write.csv(Ra_VWC2_Oct24_pwr$df, paste0(dir_pwr, "Ra_VWC20_Oct24.csv"), row.names = F)

# extracting diurnal angle & power patterns from heatmap
wc_Ra_Oct24_diurnal_summary <- wc_result_summary(GPP_SR_Oct24, "Ra", wc_Ra_Oct24)
wc_Ra_Ts20_Oct24_diurnal <- coh_result(scale_pad(GPP_SR_Oct24, "Ra", "TS_1_2_1"), wc_Ra_Ts2_Oct24)
wc_Ra_VWC20_Oct24_diurnal <- coh_result(scale_pad(GPP_SR_Oct24, "Ra", "SWC_1_2_1"), wc_Ra_VWC2_Oct24)
# save as csv
save_to_csv(wc_Ra_Oct24_diurnal_summary, dir_diurnal, "wc_Ra_Oct24_")
write.csv(wc_Ra_Ts20_Oct24_diurnal, paste0(dir_diurnal,"wc_Ra_Oct24_Ts20.csv"), row.names = F)
write.csv(wc_Ra_VWC20_Oct24_diurnal, paste0(dir_diurnal,"wc_Ra_Oct24_VWC20.csv"), row.names = F)

# viz
ggarrange(
  wc_lag_img(wc_Ra_Oct24_diurnal_summary$GPP, "GPP"),
  wc_lag_img(wc_Ra_Oct24_diurnal_summary$PAR, "PAR"),
  wc_lag_img(wc_Ra_Oct24_diurnal_summary$Ts, "Ts"),
  wc_lag_img(wc_Ra_Oct24_diurnal_summary$VWC, "VWC"),
  ncol = 2, nrow = 2, common.legend = T
)

```

```{r WT Oct2024}
wt_Rh_Oct24 <- 
  analyze.wavelet(wc_Rh_Oct24$PAR$series,"Rh", 
                  loess.span = 0,     ## no detrending required
                  dt = 1/24,          ## defines the time unit --- there are 24 slots in a day! Time unit = 1 day
                  dj = 1/100,         ## reOctlution along period axis
                  lowerPeriod = 1/4,  ## 6 hour
                  upperPeriod = 64,  ## 128 days
                  make.pval = T,      ## check for significance
                  n.sim = 100)         ## 10 white-noise simulations used for assessing significance )

Rh_Oct24_pwr <- wt_avg(wt_Rh_Oct24, 25)
write.csv(Rh_Oct24_pwr$df, paste0(dir_pwr, "Rh_Oct24.csv"), row.names = F)

wt.image(wt_Rh_Oct24,  
         legend.params = 
             list(
                 lab = "wavelet power levels", # axis label
                 lab.line = 1.5, # distance between bar label and axis label
                 mar = 5
             ), 
         lvl = 0.3,
         timelab = "",
         show.date = T, 
         date.format = "%Y-%m-%d %H:%M:%S",
         periodlab = "period (days)",
         main = "Rh"
)

# Ra
wt_Ra_Oct24 <- 
  analyze.wavelet(wc_Ra_Oct24$PAR$series,"Ra", 
                  loess.span = 0,     ## no detrending required
                  dt = 1/24,          ## defines the time unit --- there are 24 slots in a day! Time unit = 1 day
                  dj = 1/100,         ## resolution along period axis
                  lowerPeriod = 1/4,  ## 6 hour
                  upperPeriod = 64,  ## 128 days
                  make.pval = T,      ## check for significance
                  n.sim = 100)         ## 10 white-noise simulations used for assessing significance )

Ra_Oct24_pwr <- wt_avg(wt_Ra_Oct24, 25)
write.csv(Ra_Oct24_pwr$df, paste0(dir_pwr, "Ra_Oct24.csv"), row.names = F)

wt.image(wt_Ra_Oct24,  
         legend.params = 
             list(
                 lab = "wavelet power levels", # axis label
                 lab.line = 1.5, # distance between bar label and axis label
                 mar = 5
             ), 
         lvl = 0.3,
         timelab = "",
         show.date = T, 
         date.format = "%Y-%m-%d %H:%M:%S",
         periodlab = "period (days)",
         main = "Ra"
)
```
