==> Codes/04_temp_pwr_peak.Rmd <==
---
title: "extracting temporal power peaks"
author: "MOEKA"
date: "2025-05-14"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(lubridate)
library(dplyr)
library(ggplot2)
library(ggpubr)
library(tidyr)
library(gridExtra)
library(purrr)
```

```{r}
filenames_raw_pwr <- 
  list.files("G:/My Drive/Research/Projects/DC_auto/Data/Result/GPP_SR_cospectral/raw_64d_one_GPP/df_temporal_pwr/", "*.csv", full.names = T)

filenames_raw_pwr_env <- 
  list.files("G:/My Drive/Research/Projects/DC_auto/Data/Result/GPP_SR_cospectral/raw_64d_one_GPP/df_temporal_pwr_env/", "*.csv", full.names = T)

filenames_res1_pwr <- 
  list.files("G:/My Drive/Research/Projects/DC_auto/Data/Result/GPP_SR_cospectral/res_64d_one_GPP_hourly_dailyQ10/df_temporal_pwr/", "*.csv", full.names = T)

filenames_res2_pwr <- 
  list.files("G:/My Drive/Research/Projects/DC_auto/Data/Result/GPP_SR_cospectral/res_64d_one_GPP_hourly_weeklyQ10/df_temporal_pwr/", "*.csv", full.names = T)
```


```{r making dfs for diurnal peaks}
temp_pwr_csv <- function(filename){
  
  df <- strsplit(filename, "/")[[1]][length(strsplit(filename, "/")[[1]])]
  
  resp <- strsplit(df, "_")[[1]][1]

  driver <- ifelse(
      is.na(strsplit(df, "_")[[1]][3]),
      NA,
      sub("\\.csv$", "", strsplit(df, "_")[[1]][2])
    )
  campaign <- ifelse(
      is.na(strsplit(df, "_")[[1]][3]),
      sub("\\.csv$", "", strsplit(df, "_")[[1]][2]),
      sub("\\.csv$", "", strsplit(df, "_")[[1]][3])
    )
    
  result <- 
    read.csv(filename) %>%
    summarise(
      pwr_P1 = Power.avg[Period == "1"],
      pwr_max = max(Power.avg, na.rm = TRUE),
      pwr_max_period = Period[which.max(Power.avg)]) %>% 
    mutate(
      df = df,
      campaign = campaign,
      resp = resp,
      driver = driver
    ) 
}

### for raw resp
raw_pwr <- purrr::map_dfr(filenames_raw_pwr, temp_pwr_csv)
raw_pwr$campaign <- 
  factor(raw_pwr$campaign, levels = c("MJ22", "MA23", "Mar24", "Apr24", "Sep24", "Oct24"))
raw_pwr <- raw_pwr[order(raw_pwr$resp, raw_pwr$driver, raw_pwr$campaign), ]
                              
# csv
write.csv(raw_pwr, "G:/My Drive/Research/Projects/DC_auto/Data/Result/GPP_SR_cospectral/raw_64d_one_GPP/temp_pwr_raw.csv", row.names = F)

### for raw env 
raw_pwr_env <- purrr::map_dfr(filenames_raw_pwr_env, temp_pwr_csv)
raw_pwr_env$campaign <- 
  factor(raw_pwr_env$campaign, levels = c("MJ22", "MA23", "Mar24", "Apr24", "Sep24", "Oct24"))
raw_pwr_env <- raw_pwr_env[order(raw_pwr_env$resp, raw_pwr_env$driver, raw_pwr_env$campaign), ]
                              
# csv
write.csv(raw_pwr_env, "G:/My Drive/Research/Projects/DC_auto/Data/Result/GPP_SR_cospectral/raw_64d_one_GPP/temp_pwr_raw_env.csv", row.names = F)

### for res1 resp
res1_pwr <- purrr::map_dfr(filenames_res1_pwr, temp_pwr_csv)
res1_pwr$campaign <- 
  factor(res1_pwr$campaign, levels = c("MJ22", "MA23", "Mar24", "Apr24", "Sep24", "Oct24"))
res1_pwr <- res1_pwr[order(res1_pwr$resp, res1_pwr$driver, res1_pwr$campaign), ]
                              
# csv
write.csv(res1_pwr, "G:/My Drive/Research/Projects/DC_auto/Data/Result/GPP_SR_cospectral/res_64d_one_GPP_hourly_dailyQ10/temp_pwr_res1.csv", row.names = F)

### for res2 resp
res2_pwr <- purrr::map_dfr(filenames_res2_pwr, temp_pwr_csv)
res2_pwr$campaign <- 
  factor(res2_pwr$campaign, levels = c("MJ22", "MA23", "Mar24", "Apr24", "Sep24", "Oct24"))
res2_pwr <- res2_pwr[order(res2_pwr$resp, res2_pwr$driver, res2_pwr$campaign), ]
                              
# csv
write.csv(res2_pwr, "G:/My Drive/Research/Projects/DC_auto/Data/Result/GPP_SR_cospectral/res_64d_one_GPP_hourly_weeklyQ10/temp_pwr_res2.csv", row.names = F)
```


```{r making plots}
temp_pwr_plot <- function(filename, file_des) {
  
  plot_list <- list()
  
  for (i in filename) {
    
    # Extract filename from path
    df <- basename(i)  # cleaner than splitting yourself

    # Extract components from filename
    split_df <- strsplit(df, "_")[[1]]
    
    resp <- split_df[1]
    driver <- ifelse(length(split_df) < 3, NA, sub("\\.csv$", "", split_df[2]))
    campaign <- ifelse(length(split_df) < 3, sub("\\.csv$", "", split_df[2]), sub("\\.csv$", "", split_df[3]))
    
    # Determine y-axis settings based on 4th file
    # 4th = MJ22
    fourth_data <- read.csv(filename[4])
    fourth_ymax <- max(fourth_data$Power.avg, na.rm = T)

    # Define dynamic y-axis
    if (fourth_ymax <= 25) {
      y_limits <- c(0, 25)
      y_breaks <- c(0, 10, 20)
    } else {
      y_limits <- c(0, 35)
      y_breaks <- c(0, 10, 20, 30)
    }
    
    # Read data
    data <- read.csv(i)
    
    # Create plot
    p1 <- 
      data %>%
      ggplot() +
        geom_line(aes(log2(Period), Power.avg)) +
        geom_point(
          data = . %>% filter(!is.na(Power.avg.sig)), 
          aes(log2(Period), Power.avg.sig), size = 1
        ) + 
        theme_classic() +
        scale_x_continuous(
          breaks = c(0, 3, 6),
          labels = c(1, 8, 64)
        ) +
        scale_y_continuous(
          breaks = y_breaks,
          limits = y_limits
        ) +
        theme(
          axis.title.x = element_blank(), 
          axis.text.x = element_text(size = 10),
          axis.title.y = element_blank(),
          axis.text.y = element_text(size = 10),
          legend.title = element_text(size = 10), 
          legend.text = element_text(size = 10),
          panel.border = element_rect(colour = "black", fill = NA, linewidth = .5)
        )
    
    plot_list[[i]] <- p1
    
    # Save to file
    file_name <- file.path(file_des, paste(resp, "_", driver, "_",campaign, ".png", sep = ""))
    png(file_name, width = 560, height = 450, res = 300)
    print(p1)
    dev.off()
  }
}



temp_raw_pwr_plt <- 
  temp_pwr_plot(
    filenames_raw_pwr, 
    "G:/My Drive/Research/Projects/DC_auto/Data/Result/GPP_SR_cospectral/raw_64d_one_GPP/img_temp_pwr/"
    )

temp_raw_pwr_env_plt <- 
  temp_pwr_plot(
    filenames_raw_pwr_env, 
    "G:/My Drive/Research/Projects/DC_auto/Data/Result/GPP_SR_cospectral/raw_64d_one_GPP/img_temp_pwr_env/"
    )


temp_res1_pwr_plt <- 
  temp_pwr_plot(
    filenames_res1_pwr,
    "G:/My Drive/Research/Projects/DC_auto/Data/Result/GPP_SR_cospectral/res_64d_one_GPP_hourly_dailyQ10/img_temp_pwr/"
    )

temp_res2_pwr_plt <- 
  temp_pwr_plot(
    filenames_res2_pwr,
    "G:/My Drive/Research/Projects/DC_auto/Data/Result/GPP_SR_cospectral/res_64d_one_GPP_hourly_weeklyQ10/img_temp_pwr/"
    )

```
