==> Codes/05_phase_diff_fig.Rmd <==
---
title: "phase_angle_diff_fig"
author: "MOEKA"
date: "2024-11-15"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(lubridate)
library(dplyr)
library(ggplot2)
library(ggpubr)
library(tidyr)
library(gridExtra)
library(purrr)
library(patchwork)
```

```{r}
setwd("G:/My Drive/Research/Projects/DC_auto/Data/Result/GPP_SR_cospectral/")

filenames_raw <- 
  list.files("raw_64d_one_GPP/df_diurnal/", "*.csv", full.names = T)

filenames_raw_env <- 
  list.files("raw_64d_one_GPP/df_diurnal_env/", "*.csv", full.names = T)

filenames_res1 <- 
  list.files("res_64d_one_GPP_hourly_dailyQ10/df_diurnal/", "*.csv", full.names = T)

filenames_res2 <- 
  list.files("res_64d_one_GPP_hourly_weeklyQ10/df_diurnal/", "*.csv", full.names = T)

# for raw
ts_raw_files <- grep("_Ts\\.csv", filenames_raw, value = T)
par_raw_files <- grep("_PAR\\.csv", filenames_raw, value = T)
gpp_raw_files <- grep("_GPP\\.csv", filenames_raw, value = T)

# for raw env
ts_raw_env_files <- grep("_Ts\\.csv", filenames_raw_env, value = T)

# for res daily
ts_res1_files <- grep("_Ts\\.csv", filenames_res1, value = T)
par_res1_files <- grep("_PAR\\.csv", filenames_res1, value = T)
gpp_res1_files <- grep("_GPP\\.csv", filenames_res1, value = T)

# for res weekly
ts_res2_files <- grep("_Ts\\.csv", filenames_res2, value = T)
par_res2_files <- grep("_PAR\\.csv", filenames_res2, value = T)
gpp_res2_files <- grep("_GPP\\.csv", filenames_res2, value = T)
```


```{r}
phase_diff_csv <- function(filename){
  
  df <- strsplit(filename, "/")[[1]][length(strsplit(filename, "/")[[1]])]
  month <- strsplit(df, "_")[[1]][3]
  resp <- strsplit(df, "_")[[1]][2]
  driver <- sub(".csv","", strsplit(df, "_")[[1]][4])
  
  result <- 
    read.csv(filename) %>%
    # Angle: NA_flag == "0" & pval_period1 < 0.1
    select(c(date, NA_flag, pval_period1, Angle)) %>% 
    mutate(day = as.Date(date)) %>% 
    group_by(day) %>% 
    summarise(angle_avg = mean(Angle, na.rm = T)) %>% 
    mutate(phase_hr = (angle_avg * 24) / (2 * pi))
  
  result %>% 
    summarise(
      Phase_hr_avg = mean(phase_hr, na.rm = T),
      Phase_hr_sd = sd(phase_hr, na.rm = T)
      ) %>% 
    mutate( 
      df = df,
      month = month,
      resp = resp,
      driver = driver
    ) 
}

# calculate
# for raw
ts_raw_phase <- purrr::map_dfr(ts_raw_files, phase_diff_csv)
par_raw_phase <- purrr::map_dfr(par_raw_files, phase_diff_csv)
gpp_raw_phase <- purrr::map_dfr(gpp_raw_files, phase_diff_csv)

# for raw env
ts_raw_env_phase <- purrr::map_dfr(ts_raw_env_files, phase_diff_csv)

# for res1
ts_res1_phase <- purrr::map_dfr(ts_res1_files, phase_diff_csv)
par_res1_phase <- purrr::map_dfr(par_res1_files, phase_diff_csv)
gpp_res1_phase <- purrr::map_dfr(gpp_res1_files, phase_diff_csv)

# for res2
ts_res2_phase <- purrr::map_dfr(ts_res2_files, phase_diff_csv)
par_res2_phase <- purrr::map_dfr(par_res2_files, phase_diff_csv)
gpp_res2_phase <- purrr::map_dfr(gpp_res2_files, phase_diff_csv)


# vars order
campaign <- c("MJ22" = "C1", "MA23" = "C2", "Mar24" = "C3", 
              "Apr24" = "C4", "Sep24" = "C5", "Oct24" = "C6")
season_mo <- c("MJ22" = "Growing", "MA23" = "Dormant", "Mar24" = "Dormant", 
              "Apr24" = "Growing", "Sep24" = "Growing", "Oct24" = "Dormant")


# combined and save
# for raw & raw env
phase_raw_combined <- 
  rbind(ts_raw_phase, par_raw_phase, gpp_raw_phase, ts_raw_env_phase) %>% 
  mutate(
    campaign = recode(month, !!!campaign),
    season = recode(month, !!!season_mo)
    ) %>% 
  arrange(resp, driver, campaign)

write.csv(phase_raw_combined, "raw_64d_one_GPP/df_lag_hr/phase_raw_combined.csv", row.names = F)


# for res1
phase_res1_combined <- 
  rbind(ts_res1_phase, par_res1_phase, gpp_res1_phase) %>% 
  mutate(
    campaign = recode(month, !!!campaign),
    season = recode(month, !!!season_mo)
    ) %>% 
  arrange(resp, driver, campaign)

write.csv(phase_res1_combined, "res_64d_one_GPP_hourly_dailyQ10/df_lag_hr/phase_res1_combined.csv", row.names = F)


# for res2
phase_res2_combined <- 
  rbind(ts_res2_phase, par_res2_phase, gpp_res2_phase) %>% 
  mutate(
    campaign = recode(month, !!!campaign),
    season = recode(month, !!!season_mo)
    ) %>% 
  arrange(resp, driver, campaign)

write.csv(phase_res2_combined, "res_64d_one_GPP_hourly_weeklyQ10/df_lag_hr/phase_res2_combined.csv", row.names = F)
```


```{r graph}

phase_diff_graph <- function(df, res, driver_var){
  
  df %>% 
  filter(resp == res & driver == driver_var) %>% 
  ggplot(aes(campaign, Phase_hr_avg, shape = season)) +
  geom_point(size = 7) + 
  geom_errorbar(
    aes(campaign,
        ymin = Phase_hr_avg - Phase_hr_sd,
        ymax = Phase_hr_avg + Phase_hr_sd,
        ),
    lwd = 1.5, width = .2
  ) +
  geom_abline(slope = 0, intercept = 0) + 
  theme_classic() + 
  ylim(-9, 9) +
  ggtitle(paste(res, "vs", driver_var)) +
  ylab("Lag (Hour)") + 
  theme(
    axis.title.x = element_blank(),
    axis.text.x = element_text(size = 15),
    axis.title.y = element_text(size = 15),
    axis.text.y = element_text(size = 15),
    axis.ticks.x = element_blank(),
    plot.title = element_text(size = 15),
    legend.title = element_blank(),
    panel.border = element_rect(colour = "black", fill = NA, linewidth = 1)
        )
}

# for raw
# run each case
p_Rh_PAR <- phase_diff_graph(phase_raw_combined, "Rh", "PAR")
p_Ra_PAR <- phase_diff_graph(phase_raw_combined, "Ra", "PAR")

p_Rh_Ts <- phase_diff_graph(phase_raw_combined, "Rh", "Ts")
p_Ra_Ts <- phase_diff_graph(phase_raw_combined, "Ra", "Ts")

p_Rh_GPP <- phase_diff_graph(phase_raw_combined, "Rh", "GPP")
p_Ra_GPP <- phase_diff_graph(phase_raw_combined, "Ra", "GPP")

# patchwork
combined_plot <- 
  (p_Rh_GPP + p_Rh_PAR + p_Rh_Ts) / (p_Ra_GPP + p_Ra_PAR + p_Ra_Ts) +
  # common legend
  plot_layout(guides = "collect") & 
  theme(
    legend.position = "bottom",
    legend.text = element_text(size = 13)) & 
  ylab(NULL)

# wrap_elements(combined_plot) +
#     labs(tag = "Lag (hour)") +
#     theme(
#         plot.tag = element_text(size = 15, angle = 90),
#         plot.tag.position = "left"
#     )

## for raw env
# run each case
p_GPP_Ts <- phase_diff_graph(phase_raw_combined, "GPP", "Ts") + ylim(0, 8)
p_PAR_Ts <- phase_diff_graph(phase_raw_combined, "PAR", "Ts") + ylim(0, 8)


# patchwork
(p_GPP_Ts + p_PAR_Ts) +
  # common legend
  plot_layout(guides = "collect") & 
  theme(
    legend.position = "bottom",
    legend.text = element_text(size = 13)) & 
  ylab(NULL)


# for res1
# run each case
p_res1_rRh_PAR <- phase_diff_graph(phase_res1_combined, "rRh", "PAR") + ylim(-10, 10)
p_res1_rRa_PAR <- phase_diff_graph(phase_res1_combined, "rRa", "PAR") + ylim(-10, 10)

p_res1_rRh_Ts <- phase_diff_graph(phase_res1_combined, "rRh", "Ts") + ylim(-10, 10)
p_res1_rRa_Ts <- phase_diff_graph(phase_res1_combined, "rRa", "Ts") + ylim(-10, 10)

p_res1_rRh_GPP <- phase_diff_graph(phase_res1_combined, "rRh", "GPP") + ylim(-10, 10)
p_res1_rRa_GPP <- phase_diff_graph(phase_res1_combined, "rRa", "GPP") + ylim(-10, 10)

# patchwork
(p_res1_rRh_GPP + p_res1_rRh_PAR + p_res1_rRh_Ts) / (p_res1_rRa_GPP + p_res1_rRa_PAR + p_res1_rRa_Ts) +
  # common legend
  plot_layout(guides = "collect") & 
  theme(
    legend.position = "bottom",
    legend.text = element_text(size = 13)) & 
  ylab(NULL)


# for res2
# run each case
p_res2_rRh_PAR <- phase_diff_graph(phase_res2_combined, "rRh", "PAR") + ylim(-10, 10)
p_res2_rRa_PAR <- phase_diff_graph(phase_res2_combined, "rRa", "PAR") + ylim(-10, 10)

p_res2_rRh_Ts <- phase_diff_graph(phase_res2_combined, "rRh", "Ts") + ylim(-10, 10)
p_res2_rRa_Ts <- phase_diff_graph(phase_res2_combined, "rRa", "Ts") + ylim(-10, 10)

p_res2_rRh_GPP <- phase_diff_graph(phase_res2_combined, "rRh", "GPP") + ylim(-10, 10)
p_res2_rRa_GPP <- phase_diff_graph(phase_res2_combined, "rRa", "GPP") + ylim(-10, 10)

# patchwork
(p_res2_rRh_GPP + p_res2_rRh_PAR + p_res2_rRh_Ts) / (p_res2_rRa_GPP + p_res2_rRa_PAR + p_res2_rRa_Ts) +
  # common legend
  plot_layout(guides = "collect") & 
  theme(
    legend.position = "bottom",
    legend.text = element_text(size = 13)) & 
  ylab(NULL)
```
